% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_measure_tbl.R
\name{assign_group}
\alias{assign_group}
\title{Measure Assignment Into Measure Group}
\usage{
assign_group(measure_tbl = measure_tbl)
}
\arguments{
\item{measure_tbl}{A measure}
}
\value{
The measure_tbl with assignment of a group
}
\description{
Assign the measure into the seven measure group according to SAS Pack
}
