% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_measure_tbl.R
\name{cols}
\alias{cols}
\alias{measure}
\title{Measure Name Filter}
\usage{
cols(
  x = sasdata,
  type = c("score", "denominator", "wt", "id"),
  groups = NULL,
  cutoff = 1L
)
}
\arguments{
\item{x}{A measure score table.}

\item{type}{"id","score", "denominator", or "wt".}

\item{groups}{A vector of measure group.}

\item{cutoff}{A integer. The measure is removed if there is less than the
cutoff number of hospital reported.}
}
\value{
A vector of measure names.
}
\description{
Output a vector of measure name from a measure score table according to the
type and the measure group.
}
