% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_measure_tbl.R
\name{create_measure_tbl}
\alias{create_measure_tbl}
\title{Measure Table Creator}
\usage{
create_measure_tbl(data_tbl = sas_data)
}
\arguments{
\item{data_tbl}{The sas data input file.}
}
\value{
A measure table with name of the measure, type (weight, denominator or
  score), the assigned group, and hospital counts, etc.
}
\description{
Create a measure table to hold the measure information for the column of the
sas input data data
}
