% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_measure_score_table.R
\name{create_weight}
\alias{create_weight}
\title{Calculate the weights from the denominator}
\usage{
create_weight(x = r_x)
}
\arguments{
\item{x}{A data frame containing the measure denominator of the hospital compare star rating.}
}
\value{
A data frame containing the weighting factor which will used to perform lvm.
}
\description{
Calculate the weights from the denominator
}
\details{
The formula:
\deqn{w_{khd} = \frac{n_{khd}}{\sum_{i=1}^{N_{kd}}(n_{khd})} * N_{kd}}{w_khd = (n_khd)/(N_kd)}
 Where:
 \itemize{
 \item \eqn{w_{khd}}{w_khd}: Weight for measure \code{k}, hospital \code{h}, measure group \code{d}
 \item \eqn{n_{khd}}{n_khd}: Denominator for hospital \code{h}, measure \code{k} in measrue group \code{d}
 \item \eqn{N_{kd}}{N_kd}:  The total number of hospitals for measure \code{k} in measure group \code{d}
 }

 \deqn{R^2 = 1 - \frac{estimated variance}{observed variance}}{R-
squared = 1 - (estimated variance)/(observed variance)}
}
\examples{
:


}
