% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_randomize_manipulation.R
\name{measure_den_one}
\alias{measure_den_one}
\title{Set The Demoninator To One.}
\usage{
measure_den_one(dat = cms_star_rating_input_2017dec)
}
\arguments{
\item{dat}{A input data frame for cms star ratings.}
}
\value{
A data frame in which the denominators are set to 1.
}
\description{
This function is for testing purpose for the latent variable model. Set the
demoninator to 1one is a way to remove the weigting from the latent variable
model.
}
