% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_randomize_manipulation.R
\name{measure_manipulate}
\alias{measure_manipulate}
\title{Measure Manipulation In The Input Data Frame For CMS Star Ratings.}
\usage{
measure_manipulate(dat, method = c("randomize", "den_one"), measures = NULL)
}
\arguments{
\item{dat}{A input data frame for CMS star ratings.}

\item{method}{A string to tell the manipulation method.}

\item{measures}{A list of measures. This is used by randomize method only.}
}
\value{
A data frame for CMS Star Ratings.
}
\description{
randomize the measure or set the denominator to one.
}
\seealso{
\code{\link{measure_randomize}} for measure randomization,
  \code{\link{measure_den_one}} for the manipulation to set the denominator
  to one.
}
