% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_randomize_manipulation.R
\name{measure_randomize}
\alias{measure_randomize}
\title{Measure Randomization}
\usage{
measure_randomize(x, dat = cms_star_rating_input_2017dec)
}
\arguments{
\item{x}{A list of measures.}

\item{dat}{A input data frame for cms star ratings.}
}
\value{
A data frame for cms star ratings.
}
\description{
Randomize one or more measures. This is for test only.
}
\examples{
\dontrun{
# To randomize the non-null values in the specified measures in the input data frame.
measure_randomize(c("hai_1","hai_2"),dat=cms_star_rating_input_2017dec)
}

}
