% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_measure_score_table.R
\name{mstbl}
\alias{mstbl}
\title{Measure Table Clean Up.}
\usage{
mstbl(x, cutoff = 100)
}
\arguments{
\item{x}{A data frame containing measure scores and their correponding
denominators. Each column is a measure or denominator.}

\item{cutoff}{The cutoff.  The measure is removed if the number of hospital
reported in the measure is less than the cutoff.}
}
\value{
A list containing four data frames: \itemize{ \item msdtbl: measure
  score and denominator table. \item mstbl_std: winsorized and standized measure score
  table. \item  wtbl: weight table. \item report indicator: 1 means the star
  will be report for the corresponding hospital and 0 means no star is
  reported for the corresponding hospital. }
}
\description{
Clean
}
