% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{out_dir}
\alias{out_dir}
\title{Output Directory}
\usage{
out_dir(x)
}
\arguments{
\item{x}{A full path directory string.}
}
\value{
A directory string.
}
\description{
Check the input string as a directory. Create a directory if it doesn't exists.
}
