% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summary_score.R
\name{reweight}
\alias{reweight}
\title{Redistribute weights when there is missing group}
\usage{
reweight(grp = rep(1, 7), wt = NULL)
}
\arguments{
\item{grp}{The data to be weighted. For example, the group score in the cms
SAS Pack star rating.}

\item{wt}{A vector of weights. Refer to line 95-102 in Sas pack #2 k-mean
clustering.}
}
\value{
The weight redistributed. the weight set to zero when the
  corresponding term is missing and the corresponding weigt is redistributed
  to other term.
}
\description{
A function to recalculate weights. the corresponding weigt is redistributed
into other group if the group is completely missed.
}
\examples{
reweight(grp=c(1,1,1,1,1,1,NA))

}
