% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{subgroup}
\alias{subgroup}
\title{Prepare the subset of the data in a measure group for CMS star rating}
\usage{
subgroup(object, groups = NULL)
}
\arguments{
\item{object}{A mstbl object,which contains the measure score table and
weight table.}

\item{groups}{A vector of group names. The default is null and it will
automatically use all of the groups in the object.}
}
\value{
A list containing a sublist for each measure group. Each has a
  stadardized measure score table and a weight table.
}
\description{
Prepare the subset of the data table and weight table in each measure group
in CMS Hospital Compare star rating. In addition, the hospital is removed if
the hospital didn't report any data in the group.
}
