% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summary_score.R
\name{sum_score}
\alias{sum_score}
\title{Calculation Of Summary Score From Group Score}
\usage{
sum_score(x, wts = NULL)
}
\arguments{
\item{x}{A data frame containing seven group scores.}

\item{wts}{A vector containing the arbitrary weighting corresponding to input
data frame.}
}
\value{
Summary scores corresponding to each row from the input data frame
}
\description{
A function to calculate the summary score from group score. The weigt of a
group score is redistributed into other group if the group is missed.
}
