% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report_indicator.R
\name{valid_report}
\alias{valid_report}
\title{Report Indicator Generator}
\usage{
valid_report(x)
}
\arguments{
\item{x}{A data frame with each measure in a column and each hospital in a
row.}
}
\value{
A vector of report indicators: 0 as not being reported, or 1 as
  getting reported.
}
\description{
Create a report indicator for each hospital according to the reporting
criteria setted by CMS.
}
\details{
Reporting criteria setted by CMS: 1. minimum measure threshold: 3 measues per
group 2. minimum group threshold:   3 group per hospital, with at least on
Outcome group (those are, Mortality, Safty of Care, Readmission)
}
