% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_measure_score_table.R
\name{winsorize}
\alias{winsorize}
\title{Winsorization Function}
\usage{
winsorize(x, min = qnorm(0.05), max = qnorm(0.95))
}
\arguments{
\item{x}{A numeric vector to be winsorized. It should be scaled and centered
to a normal distribution.}

\item{min}{The left side cutoff}

\item{max}{The right side cutoff}
}
\value{
A winsorized vector
}
\description{
Winsorize a vector according to the cutoff parameters specified by min and
max.
}
\examples{
out <- winsorize(x = rnorm(1000), min=qnorm(0.05), max = qnorm(0.95))
hist(out,breaks=20)

}
